Create Python_LiveProject
The AppBuilder9000 Overview

This project was built using the Django framework. This project is the interactive website for managing one's collections of things related to various hobbies, as well as API and Data Scraped content for those hobbies.  
Each developer contributes to the overall project by adding a new App that utilizes API's/Data Scrapping to receive content for the hobby of their choice.
Contributing app = Brother_EDGAR
Brother EDGAR is an app designed to utilize API functionality to gain better insight into the financial/Stock market.  The name EDGAR is based on SEC.gov's search engine which stands for Electronic Data Gathering  Analysis and Retrieval system.